requirements.txt
python-telegram-bot
python-dotenv
pandas
matplotlib
requests
numpy
imgkit
pdfkit
pymongo
schedule
